Generate GraphQL schema from dataset dimensions.
Generate a GraphQL schema for the dimensions in a dataset. Reference
area and reference period dimensions are mapped to custom scalars which
convert a given string value to the ref area/period URI. Other
dimensions are mapped to custom enum types.

Add query evaluator which allows top-level properties (URI, label,
description) of the dataset to be queried. The observation query
currently returns an empty list.

The server namespace contains functions to create and host a pedestal
server through which queries can be executed.