Extract dimension and measure type bindings in the same way.
Extract and convert the dimension and measure type fields from a set
of observation bindings in the same way: obtain the query variable
name, retreive the binding value and convert it to a GraphQL value
appropriate for the declared schema type. This is done by adding
->query-var-name and result-binding->value functions to both
dimension and measure-type maps.