Add dataset query root.
Add query root which allows datasets to be filtered by the dimensions
they contain. Clients can specify a filter of the form {and: [uris]
or: [uris]}, returned datasets must contain all dimensions specified
in the and list and at least one from the or list. If either list is
empty it will not be used to filter any datasets.