Generate aggregation schema for each dataset.
Add an aggregations field for each dataset observations type which
allow measure type aggregations. Only measure type with numeric values
can be aggregated - the type of the measure type is inferred from the
associated value on a random observation within the dataset (all
values for the meausre are assumed to have the same type).