Require pagination for observation results.
Require pagination for matching observation results instead of returning
all matches. Add a page field to the observations schema and move the
:first and :after arguments from observations. If these are not provided,
defaults of 10 and 0 are used respectively.

The page field contains the list of matching observations along with the
next page and total number of results returned.

LIMIT and OFFSET parameters are only required for paging queries so
remove it from the top-level observations sparql field.