Allow ordering of matching observations.
Update the schema to allow the returned observations to be ordered
by multiple fields. The observations field has two new arguments:
order which has the list of fields to sort by in priority order,
and an optional order_spec parameter which specifies whether to apply
the sort for the field in ascending or descending order. The default
for all fields is descending order.

Apply the sort for each measure/dimension to the generated SPARQL
query for observations.