Ensure enum cases have different names.
Dimension value labels are not guaranteed to map to unique enum
case names. Ensure uniqueness by appending a number to the end of
the names of dimension values which would normally map to the same
name e.g. if there are multiple values which would map to the enum
name NAME, map them instead to NAME_1, NAME_2, etc.

Point the scotland SPARQL repository to the staging endpoint instead
of the live site.