Add nested properties to the result schema for refPeriod dimensions.
Issue #72 - Allow measures and dimensions to define the variables
to bind in the observations query and construct the GraphQL value
returned to the client. This allows dimensions to be mapped to more
than a single value within the GraphQL schema.

Change the observation schema type of the refPeriod dimension to
include four nested properties - the period URI, label and interval
start and end times.