Make endpoint parameter required.
Issue #98 - Make the endpoint a required parameter and raise an error
if not specified. Allow a source file or directory containing RDF
data to be used as the endpoint instead of requiring a remote SPARQL
endpoint. Update the README to document the endpoint parameter and
include an example for using the development test data.