Fetch dataset title according to specified language.
Derive the dataset schema name from the string literal (i.e. without
a language code) asscoiated with the dataset title predicate. Make
the dataset title and optional property which may have an associated
language code. Fetch this title along with the rest of the dataset
properties.

Fetch the dataset title in the specified language for each
custom dataset schema resolver.