Fix validation queries for execution within rdf-validator.
Move the comments for the dimension validation queries to the end of
of the file. Comments before the query cause sesame to infer the
wrong query type (i.e. a graph query instead of a tuple query) which
results in the wrong accept headers being sent to the remote SPARQL
endpoint.